+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in the last 3 major versions of the postg
 
 ## Unreleased
 
+- Correctly configure postgres-common on ubuntu hosts, fixes #596
+
 ## v7.1.6 (2020-02-20)
 
 - Remove unnecessary nil default in resource properties